+++ package.json
@@ -11,7 +11,6 @@
     "format": "prettier --write --plugin prettier-plugin-svelte ."
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tsconfig/svelte": "^5.0.4",
     "@types/node": "^22.14.0",
     "cpx": "^1.5.0",
@@ -27,7 +26,7 @@
   "dependencies": {
     "@sveltejs/adapter-auto": "^6.0.0",
     "@sveltejs/kit": "^2.20.4",
-    "svelte-routing": "^2.13.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "svelte-transition": "^0.0.17"
   }
 }
