--- src/lib/components/RouteAuthGuard.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-  import { user } from '$lib/store/user.store'
-
-  let { children } = $props()
-</script>
-
-{#if $user}
-  {@render children()}
-{/if}
