+++ src/lib/store/user.store.ts
@@ -1,6 +1,15 @@
 import { writable } from 'svelte/store'
-import type { User } from 'types/user.types'
+import { browser } from '$app/environment'
+import type { User } from '$types/user.types'
 
-const storage = localStorage.getItem('name')
+const stored = browser ? localStorage.getItem('name') : null
+const initial = stored ? { name: stored } : null
 
-export const user = writable<User>(storage ? { name: storage } : null)
+export const user = writable<User | null>(initial)
+
+user.subscribe((value) => {
+  if (browser) {
+    if (value) localStorage.setItem('name', value.name)
+    else localStorage.removeItem('name')
+  }
+})
