--- src/pages/private/Logout/Logout.scss
@@ -1,14 +1,0 @@
-#logout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-  padding: 50px 20px;
-
-  .title {
-    font-size: 2rem;
-    color: white;
-    text-shadow: 2px 2px 2px rgba(0, 0, 0, 0.527);
-    margin-bottom: 50px;
-  }
-}
